LandsatExplorer

Documentation for LandsatExplorer.

LandsatExplorer.BoundingBox

LandsatExplorer.Point

LandsatExplorer.acquisition_filter

LandsatExplorer.api_authenticate

LandsatExplorer.api_request

LandsatExplorer.authenticate_download

LandsatExplorer.cloud_filter

LandsatExplorer.download_scene

LandsatExplorer.get_data_id

LandsatExplorer.get_dataset

LandsatExplorer.get_entity_id

LandsatExplorer.guess_dataset

LandsatExplorer.is_entity_id

LandsatExplorer.is_scene_id

LandsatExplorer.logged_in

LandsatExplorer.logout

LandsatExplorer.parse_entity_id

LandsatExplorer.parse_scene_id

LandsatExplorer.search

LandsatExplorer.spatial_filter

LandsatExplorer.BoundingBox — Type

BoundingBox(ul, lr)

Construct a bounding box defined by the corners ul and lr.

All coordinates should be provided in latitude and longitude.

Parameters

ul: The upper-left corner of the box as a Tuple{T,T} of latitude and longitude.

lr: The lower-right corner of the box as a Tuple{T,T} of latitude and longitude.

Example

bb = BoundingBox((52.1, -114.4), (51.9, -114.1))

LandsatExplorer.Point — Type

Point(lat, lon)

Construct a point located at the provided latitude and longitude.

Parameters

lat: The latitude of the point.

lon: The longitude of the point.

Example

p = Point(52.0, -114.25)

LandsatExplorer.acquisition_filter — Method

Filter by acquisition date

LandsatExplorer.api_authenticate — Method

Authenticate with the m2m API and return an access token

LandsatExplorer.api_request — Method

Send a get request to the m2m API

LandsatExplorer.authenticate_download — Method

Generate authentication cookies for downloading scenes

LandsatExplorer.cloud_filter — Method

Filter by cloud coverage

LandsatExplorer.download_scene — Function

download_scene(scene, dir=pwd(); log_progress=true, unpack=false)

Download a Landsat scene and save to the given directory.

Parameters

scene: The display name or entity ID of the scene to be downloaded.

dir: The directory in which to save the downloaded scene.

Keywords

unpack: If true, unpacks and deletes the downloaded tar file (default = false).

log_progress: If true, logs the download progress at 1-second intervals (default = true).

Returns

The path to the downloaded file(s).

LandsatExplorer.get_data_id — Method

Return the data id to download a scene belonging to the given dataset

LandsatExplorer.get_dataset — Method

Return the dataset string for the provided satellite and processing level

LandsatExplorer.get_entity_id — Method

get_entity_id(scene)

Lookup the entity ID for the given Landsat scene.

Example

julia> LE.get_entity_id("LC08_L2SP_043024_20200802_20200914_02_T1")
"LC80430242020215LGN00"

LandsatExplorer.guess_dataset — Method

Guess the dataset string for the given identifier

LandsatExplorer.is_entity_id — Method

Returns true if string matches format for entity id

LandsatExplorer.is_scene_id — Method

Returns true if string matches format for scene id

LandsatExplorer.logged_in — Method

Returns true if logged in to the download API

LandsatExplorer.logout — Method

Logout of earthexplorer; this is necessary to prevent authentication errors down the line

LandsatExplorer.parse_entity_id — Method

Parse metadata from Landsat entity ID

LandsatExplorer.parse_scene_id — Method

Parse metadata from Landsat scene ID

LandsatExplorer.search — Method

search(satellite, level; dates=nothing, clouds=nothing, geom=nothing, max_results=100)

Search for Landsat scenes belonging to the given satellite and processing level.

Parameters

satellite: One of "LANDSAT5", "LANDSAT7", or "LANDSAT8", or "LANDSAT9".

level: The processing level; either 1 or 2.

Keywords

dates: The date range for image acquisition. Should be a tuple of DateTime objects.

clouds: The maximum allowable cloud cover as a percentage.

geom: A geometry specifying the region of interest. Can be a Point, BoundingBox, or any other GeoInterface compatible geometry.

max_results: The maximum number of results to return (default = 100).

Returns

A DataFrame with the columns :Name, :AcquisitionDate, :PublicationDate, :CloudCover, and :Id.

Example

julia> p = Point(52.0, -114.2);

julia> dates = (DateTime(2020, 8, 1), DateTime(2020, 9, 1));

julia> search("LANDSAT_8", 2, dates=dates, geom=p, clouds=21)
3×5 DataFrame
 Row │ Name                               AcquisitionDate      Pub ⋯
     │ String                             String               Str ⋯
─────┼──────────────────────────────────────────────────────────────
   1 │ LC08_L2SP_043024_20200818_202008…  2020-08-18 00:00:00  202 ⋯
   2 │ LC08_L2SP_042024_20200811_202009…  2020-08-11 00:00:00  202
   3 │ LC08_L2SP_043024_20200802_202009…  2020-08-02 00:00:00  202
                                                   3 columns omitted

LandsatExplorer.spatial_filter — Method

Filter by geographic location